refactor(lens): split views.py into a views/ package
Behavior-preserving reorganization of the 2490-line lens/views.py into a
views/ package, following the accounts/views convention (submodules +
__init__.py re-exporting the public names with __all__). External
imports (lens/urls.py, tests) are unchanged.

Modules: base (shared viewsets, EventStreamRenderer, stream/run helpers,
LensNodeAuthMixin), credentials, lensnodes, assistants, sessions (+runs,
attachments, deliverable download, run_stream_view), datasources, skills
(+mcp), global_settings, gateway, admin_runs, shares.

Dedup: the identical _authenticate_lensnode method, previously copied in
three gateway views, is now a single LensNodeAuthMixin.

Test mock.patch targets were repointed to the modules where the patched
names are now looked up (behavior unchanged).
