+++ backend/lens/views/__init__.py
@@ -1,0 +1,57 @@
+"""Views for LensNode workers, assistants, sessions, and sharing."""
+
+from .admin_runs import (
+    AdminRunDetailView,
+    AdminRunListView,
+    _admin_run_step_counts,
+)
+from .assistants import AssistantViewSet, PublicAssistantView
+from .credentials import DataSourceCredentialViewSet
+from .datasources import DataSourceViewSet
+from .gateway import (
+    LensNodeAIGatewayView,
+    LensNodeDeliverableUploadView,
+    LensNodeSkillPackageView,
+)
+from .global_settings import GlobalSettingViewSet
+from .lensnodes import LensNodeViewSet
+from .sessions import (
+    LensAttachmentView,
+    RunOutputFileDownloadView,
+    RunViewSet,
+    SessionViewSet,
+    run_stream_view,
+)
+from .shares import (
+    AdminSharedQAViewSet,
+    PublicSharedQAListView,
+    PublicSharedQAView,
+    SharedQAViewSet,
+)
+from .skills import MCPServerViewSet, SkillViewSet
+
+__all__ = [
+    "AdminRunDetailView",
+    "AdminRunListView",
+    "AdminSharedQAViewSet",
+    "AssistantViewSet",
+    "DataSourceCredentialViewSet",
+    "DataSourceViewSet",
+    "GlobalSettingViewSet",
+    "LensAttachmentView",
+    "LensNodeAIGatewayView",
+    "LensNodeDeliverableUploadView",
+    "LensNodeSkillPackageView",
+    "LensNodeViewSet",
+    "MCPServerViewSet",
+    "PublicAssistantView",
+    "PublicSharedQAListView",
+    "PublicSharedQAView",
+    "RunOutputFileDownloadView",
+    "RunViewSet",
+    "SessionViewSet",
+    "SharedQAViewSet",
+    "SkillViewSet",
+    "_admin_run_step_counts",
+    "run_stream_view",
+]
